[spirv] Factor out Alignment and Size calculator.
